requirements.txt
gradio
requests
fastapi
httpx
pydantic
PyYAML
